add condition helpers to Deleter and Updater
